Refactor GCS data module into subpackage
`services/gcs/data.py` is now `services/gcs/data/` in the repo, broken
into the collection types and the role type.

To simplify further and prepare for additional GCS-related types to be
added, several helpers are moved from these payload classes to the
base `utils.PayloadWrapper` class. This makes them usable for all
payload building helpers.
