+++ src/globus_sdk/services/gcs/data/__init__.py
@@ -1,0 +1,13 @@
+from .collections import (
+    CollectionDocument,
+    GuestCollectionDocument,
+    MappedCollectionDocument,
+)
+from .role import GCSRoleDocument
+
+__all__ = (
+    "MappedCollectionDocument",
+    "GuestCollectionDocument",
+    "CollectionDocument",
+    "GCSRoleDocument",
+)
