+++ src/globus_sdk/services/gcs/data/collections.py
@@ -1,5 +1,5 @@
 import abc
-from typing import Any, Callable, Dict, Iterable, Optional, Tuple
+from typing import Any, Dict, Iterable, Optional, Tuple
 
 from globus_sdk import utils
 from globus_sdk.types import UUIDLike
@@ -189,24 +189,6 @@ class CollectionDocument(utils.PayloadWrapper, abc.ABC):
     def _ensure_datatype(self) -> None:
         if "DATA_TYPE" not in self:
             self["DATA_TYPE"] = f"collection#{self._deduce_datatype_version()}"
-
-    def _set_value(
-        self, key: str, val: Any, callback: Optional[Callable[[Any], Any]] = None
-    ) -> None:
-        if val is not None:
-            self[key] = callback(val) if callback else val
-
-    def _set_optstrs(self, **kwargs: Any) -> None:
-        for k, v in kwargs.items():
-            self._set_value(k, v, callback=str)
-
-    def _set_optstrlists(self, **kwargs: Optional[Iterable[Any]]) -> None:
-        for k, v in kwargs.items():
-            self._set_value(k, v, callback=lambda x: list(utils.safe_strseq_iter(x)))
-
-    def _set_optbools(self, **kwargs: Optional[bool]) -> None:
-        for k, v in kwargs.items():
-            self._set_value(k, v, callback=bool)
 
 
 class MappedCollectionDocument(CollectionDocument):
@@ -441,51 +423,3 @@ class GuestCollectionDocument(CollectionDocument):
             user_credential_id=user_credential_id,
         )
         self._ensure_datatype()
-
-
-class GCSRoleDocument(utils.PayloadWrapper):
-    """
-    Convenience class for constructing a Role document
-    to use as the `data` parameter to `create_role`
-
-    :param DATA_TYPE: Versioned document type.
-    :type DATA_TYPE: str
-    :param collection: Collection ID for the collection the role will apply to.
-        This value is omitted when creating an endpoint
-        role or when creating role definitions when creating collections.
-    :type collection: str or UUID, optional
-    :param principal: Auth identity or group id URN. Should be in the format
-        urn:globus:auth:[identity|group]:{uuid of identity or group}
-    :type principal: str, optional
-    :param role: Role assigned to the principal. Known values are owner,
-        administrator, access_manager, activity_manager, and activity_monitor
-    :type role: str, optional
-    """
-
-    def _set_value(
-        self, key: str, val: Any, callback: Optional[Callable[[Any], Any]] = None
-    ) -> None:
-        if val is not None:
-            self[key] = callback(val) if callback else val
-
-    def _set_optstrs(self, **kwargs: Any) -> None:
-        for k, v in kwargs.items():
-            self._set_value(k, v, callback=str)
-
-    def __init__(
-        self,
-        DATA_TYPE: str = "role#1.0.0",
-        collection: Optional[UUIDLike] = None,
-        principal: Optional[str] = None,
-        role: Optional[str] = None,
-        additional_fields: Optional[Dict[str, Any]] = None,
-    ) -> None:
-        super().__init__()
-        self._set_optstrs(
-            DATA_TYPE=DATA_TYPE,
-            collection=collection,
-            principal=principal,
-            role=role,
-        )
-        if additional_fields is not None:
-            self.update(additional_fields)
